requirements.txt
pyyaml==5.3.1
requests==2.23.0
pre-commit==2.1.0
bandit==1.6.2
black==19.10b0
wheel==0.34.2
twine==3.1.1
pytest-mock==3.0.0
mock==4.0.2
coverage==5.1
